Zenflows.VF.ProcessGroup: introduce ProcessGroups
Introducing ProcessGroups.  They're an extention to VF standard,
but put under the VF namespace because of convence reasons.

They're similar to filesystems without links (Unix ln(1)), in the
sense that a ProcessGroup is a Directory and a Process is a File.
The only restriction is that Directories can only contain Files XOR
other Directories at a time.