mig: modify zf_file and add zf_file_join tables
The zf_file table is now used for storing files, their hashes (used as
the primary key), sizes, and contents (named "bin").

The zf_file_join links zf_file to agents, resources, intents, and such records
that require files (images in this case) with some metadata such as name,
extension, mime type, and so on.