file,file/{domain,type,join}: reflect the db schema changes
Reflect the recent DB schema changes and add some useful utilities to help ease
work on migration to the new changes with records that require files.

* preload_gql/3 is used for preloading files of a record under the provided key
(which must point to a virtual field of array of maps) using the provided
foreign key.
* multi_insert/4 is used for inserting files of a record under the provided key
(which must point to a virtual field of array of maps) using the provided
foreign key.
* multi_update/4 is used for updating files of a record under the provided key
(which must point to a virtual field of array of maps) using the provided
foreign key.